Refactor SeedData folder
