+++ ASP.NET-Core-Webapp/Properties/launchSettings.json
@@ -20,7 +20,7 @@
       "launchBrowser": true,
       "environmentVariables": {
         "SQLCONNSTR_environmentString": "Server=(localdb)\\MSSQLLocalDB;Database=LevelUpDbProduction;Trusted_Connection=True;MultipleActiveResultSets=true",
-        "ASPNETCORE_ENVIRONMENT": "Production"
+        "ASPNETCORE_ENVIRONMENT": "Development"
       },
       "applicationUrl": "http://localhost:5001/"
     }
