+++ ASP.NET-Core-Webapp/SeedData/Seed.cs
@@ -1,131 +1,153 @@
-﻿using Newtonsoft.Json;
+﻿using ASP.NET_Core_Webapp.Data;
+using ASP.NET_Core_Webapp.Entities;
+using Microsoft.Extensions.Configuration;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
-using System.Data;
 using System.IO;
 using System.Linq;
-using System.Threading.Tasks;
-using ASP.NET_Core_Webapp.Entities;
-using ASP.NET_Core_Webapp.Data;
 
-namespace ASP.NET_Core_Webapp.Entities
+namespace ASP.NET_Core_Webapp.SeedData
 {
     public class Seed
     {
         public ApplicationContext ApplicationContext { get; set; }
-        public string Json { get; set; }
         public string Json2 { get; set; }
-        public DataSet DataSet { get; set; }
-        public DataSet SeedDataSet { get; set; }
-        public DataTable DataTable { get; set; }
-        public DataTable Library { get; set; }
-        public DataTable Users { get; set; }
-        public DataTable Archetypes { get; set; }
-        public DataTable Pitches { get; set; }
         public StreamReader StreamReader { get; set; }
         public ApplicationContext DataBase { get; set; }
+        public SeedObject SeedObject { get; set; }
+        public IConfiguration Configuration { get; set; }
 
-        public Seed(ApplicationContext applicationContext)
+        public Seed(ApplicationContext applicationContext, IConfiguration configuration)
         {
+            this.Configuration = configuration;
             this.DataBase = applicationContext;
-            this.StreamReader = new StreamReader(@"C:\Users\laszl\Documents\Programozás\greenfox\megem_project\megalotis-emerald\ASP.NET-Core-Webapp\SeedData\SeedData.json");
+            this.StreamReader = new StreamReader(Configuration["SeedLocation:FileLocation"]);
             Json2 = StreamReader.ReadToEnd();
-            Json = @"{
-          'Table1': [
-            {
-              'id': 0,
-              'item': 'item 0'
-            },
-            {
-              'id': 1,
-              'item': 'item 1'
-            }
-          ]
-        }";
-            this.DataSet = JsonConvert.DeserializeObject<DataSet>(Json);
-            DataTable = this.DataSet.Tables["Table1"];
-
-            this.SeedDataSet = JsonConvert.DeserializeObject<DataSet>(Json2);
-            Library = this.SeedDataSet.Tables["library"];
-            Users = this.SeedDataSet.Tables["users"];
-            Archetypes = this.SeedDataSet.Tables["archetypes"];
-            Pitches = this.SeedDataSet.Tables["pitches"];
+            this.SeedObject = JsonConvert.DeserializeObject<SeedObject>(Json2);
         }
 
         public Seed()
         {
-            this.StreamReader = new StreamReader(@"C:\Users\laszl\Documents\Programozás\greenfox\megem_project\megalotis-emerald\ASP.NET-Core-Webapp\SeedData\SeedData.json");
+            this.StreamReader = new StreamReader(Configuration["SeedLocation:FileLocation"]);
             Json2 = StreamReader.ReadToEnd();
-            Json = @"{
-          'Table1': [
-            {
-              'id': 0,
-              'item': 'item 0'
-            },
-            {
-              'id': 1,
-              'item': 'item 1'
-            }
-          ]
-        }";
-            this.DataSet = JsonConvert.DeserializeObject<DataSet>(Json);
-            DataTable = this.DataSet.Tables["Table1"];
-
-            this.SeedDataSet = JsonConvert.DeserializeObject<DataSet>(Json2);
-            Library = this.SeedDataSet.Tables["library"];
-            Users = this.SeedDataSet.Tables["users"];
-            Archetypes = this.SeedDataSet.Tables["archetypes"];
-            Pitches = this.SeedDataSet.Tables["pitches"];
+            this.SeedObject = JsonConvert.DeserializeObject<SeedObject>(Json2);
         }
 
-        public void writeRowCount()
+        public void FillDatabaseFromObject()
         {
-            Console.WriteLine(DataTable.Rows.Count);
-            // 2
-        }
+            var userList = SeedObject.users;
 
-        //public void writeAllRows()
-        //{
-        //    foreach (DataRow row in DataTable.Rows)
-        //    {
-        //        Console.WriteLine(row["id"] + " - " + row["item"]);
-        //    }
-        //    // 0 - item 0
-        //    // 1 - item 1
-        //    foreach (DataRow row in Library.Rows)
-        //    {
-        //        Console.WriteLine(row[0] + " " + row[1] + " " + row[2] +" "+row[3]);
-        //    }
-
-        //    foreach (DataRow row in Users.Rows)
-        //    {
-        //        Console.WriteLine(row[0] + " " + row[1] + " " + row[2] + " " + row[3] + " " + row[4]);
-        //    }
-        //    Console.WriteLine(Json2);
-        //}
-
-        public void FillDatabase()
-        {
-            //levels is a datatable
-
-            foreach (DataRow row in Library.Rows)
+            for (int i = 0; i < userList.Length; i++)
             {
-                var listOfLevels = row.Table.Columns.Cast<DataColumn>().Select(x => x.ColumnName).ToList();
+                Entities.User userToAdd = new Entities.User() { Name = userList[i].name, Picture = userList[i].pic, OpenId = userList[i].tokenAuth };
+                DataBase.Add(userToAdd);
+                DataBase.SaveChanges();
+            }
 
-                Badge badge = new Badge() { Version = row[0].ToString(), Name = row[1].ToString(), Tag = row[2].ToString() };
-                DataBase.Add(badge);
-            }
-            foreach (DataRow row in Users.Rows)
+            var badgesList = SeedObject.library;
+            for (int i = 0; i < badgesList.Length; i++)
             {
-                User user = new User() { Name = row[0].ToString(), Picture = row[2].ToString() };
-                DataBase.Add(user);
+                List<BadgeLevel> levelList = new List<BadgeLevel>();
+                foreach (var item in badgesList[i].levels)
+                {
+                    BadgeLevel badgeLevel = new BadgeLevel() { Description = item.description, Level = item.level };
+                    levelList.Add(badgeLevel);
+                }
+                Entities.Badge badgeToAdd = new Entities.Badge() { Name = badgesList[i].name, Tag = badgesList[i].tag, Version = badgesList[i].version, Levels = levelList };
+                DataBase.Add(badgeToAdd);
+                DataBase.SaveChanges();
+
+                var holders = badgesList[i].levels;
+                foreach (var item in holders)
+                {
+                    foreach (var user in item.holders)
+                    {
+                        Entities.User userToAddFromBadges = new Entities.User();
+                        userToAddFromBadges.Name = user;
+                        if (DataBase.Users.Where(x => x.Name == user).FirstOrDefault() == null)
+                        {
+                            DataBase.Add(userToAddFromBadges);
+                            DataBase.SaveChanges();
+                        }
+                        DataBase.SaveChanges();
+                    }
+                }
             }
-            foreach (DataRow row in Pitches.Rows)
-            {
-                Pitch pitch = new Pitch() { TimeStamp = DateTime.Parse(row[6].ToString()), PitchMessage = row[4].ToString(), PitchedLevel = int.Parse(row[3].ToString()) };
-                DataBase.Add(pitch);
-            }
+
             DataBase.SaveChanges();
+
+            var pitchList = SeedObject.pitches;
+
+            for (int i = 0; i < pitchList.Length; i++)
+            {
+                Entities.Pitch pitchToAdd = new Entities.Pitch() { };
+                // Add user:
+                Entities.User userToAdd = new Entities.User();
+                userToAdd = DataBase.Users.Where(x => x.Name == pitchList[i].username).FirstOrDefault();
+                // Add reviews:
+                foreach (var holderReview in pitchList[i].holders)
+                {
+                    Review review = new Review(holderReview.message, holderReview.pitchStatus);
+                    pitchToAdd.Holders.Add(review);
+                    DataBase.SaveChanges();
+
+                    review.User = DataBase.Users.Where(x => x.Name == holderReview.name).FirstOrDefault();
+                    review.Pitch = DataBase.Pitches.Where(x => x.PitchMessage == holderReview.message).FirstOrDefault();
+
+                    DataBase.Reviews.Add(review);
+                    DataBase.SaveChanges();
+                }
+
+                if (userToAdd == null)
+                {
+                    Entities.User newUser = new Entities.User() { Name = pitchList[i].username };
+                    //DataBase.Add(newUser);
+                    pitchToAdd.User = newUser;
+                    DataBase.SaveChanges();
+                }
+                else
+                {
+                    pitchToAdd.User = userToAdd;
+                    DataBase.SaveChanges();
+                }
+
+                Entities.Badge badgeToAdd;
+                badgeToAdd = DataBase.Badges.Where(x => x.Name == pitchList[i].badgeName).FirstOrDefault();
+                pitchToAdd.Badge = badgeToAdd;
+
+                pitchToAdd.TimeStamp = DateTime.ParseExact(pitchList[i].timestamp, "yyyy-MM-dd HH:mm:ss", System.Globalization.CultureInfo.InvariantCulture);
+                pitchToAdd.PitchMessage = pitchList[i].pitchMessage;
+                pitchToAdd.Badge = DataBase.Badges.Where(x => x.Name == pitchList[i].badgeName).FirstOrDefault();
+                pitchToAdd.BadgeLevel = DataBase.BadgeLevels.Where(x => x.Badge.Name == pitchList[i].badgeName).FirstOrDefault();
+                DataBase.Add(pitchToAdd);
+
+                DataBase.SaveChanges();
+
+                var badgesList2 = SeedObject.library;
+
+                foreach (var badge in badgesList2)
+                {
+                    foreach (var level in badge.levels)
+                    {
+                        foreach (var user in level.holders)
+                        {
+                            UserLevel userLevel = new UserLevel();
+
+                            BadgeLevel badgeLevelNew = DataBase.BadgeLevels.Where(x => x.Description == level.description).FirstOrDefault();
+
+                            userLevel.User = DataBase.Users.Where(x => x.Name == user).FirstOrDefault();
+                            userLevel.Badgelevel = badgeLevelNew;
+
+                            if (DataBase.UserLevels.Count(x => x.BadgeLevelId == userLevel.Badgelevel.BadgeLevelId && x.UserId == userLevel.User.UserId) == 0)
+                            {
+                                DataBase.Add(userLevel);
+                                DataBase.SaveChanges();
+                            }
+                        }
+                    }
+                }
+            }
         }
     }
 }
